#991: refactor favorites and card
